requirements.txt
apscheduler==3.10.4
psycopg[binary]==3.2.2
gspread==6.1.2
google-auth==2.35.0
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.1
tzlocal==5.2

